Use keepalive-workflow at the end of run
This adds [keepalive-workflow](https://github.com/gautamkrishnar/keepalive-workflow)
to the end of the scheduled task. After 60 days
with no activity on the master branch, Github
considers a repo "inactive" and will stop all
scheduled tasks from running. This  actions adds an empty commit every 50 days to the
master branch of the repo to keep the repo
"active"  and make sure the scheduled task
continues to run after the 60 day period. It's a
bit of a blunt instrument, but so far, there
doesn't seem to be an alternative.