Fix expectation to use options hash
Updating rspec to 3.12.0 brings with it an update to rspec-mocks to
3.12.3.

rspec-mocks since version 3.12.0 fixes a behaviour where `with` was
allowing keyword arguments to match with options hashes for ruby >= 3.

Fix the spec by setting up the expectation to receive an options hash,
which is what `GoldenRetriever::Deal#initialize` accepts.